scathat-ai-model/model-aggregator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6

# Optional: for enhanced functionality
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Blockchain integration
web3==6.11.0
eth-account==0.9.0
python-dotenv==1.0.0